+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   submission script. E.g. "--mail=BEGIN,END,FAIL"
 - FIX: numpy 2.0 compatibility
 - FIX: boolean values are now correctly stored in the sqlite database
+- FIX: When we try to open an h5 file, we now only put the program into a
+  waiting state if we catch the `BlockedIOError`. All other `OSError` are raised
+  immediately.
 
 ## v0.7.0 (12Jun24)